requirements.txt
fastapi
uvicorn
motor
beanie
pymongo
elevenlabs
groq
pyjwt
python-dotenv
pydantic
email-validator
dnspython
requests
httpx
asgiref
